feat(gallery-component): implementation
...


...


...
