+++ src/app/components/common/gallery/gallery.component.scss
@@ -1,0 +1,72 @@
+wm-gallery {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  align-content: flex-start;
+  align-items: flex-start;
+  position: relative;
+  > img {
+    flex: 0 0 auto;
+    align-self: auto;
+    border-radius: 15px;
+    padding: 10px;
+    object-fit: cover;
+    height: 150px;
+    width: 50%;
+    border-radius: 15px;
+    cursor: pointer;
+  }
+  webmapp-image-modal {
+    width: 50%;
+    padding: 10px;
+    max-width: fit-content;
+    left: 70px;
+  }
+  ion-slides {
+    width: -webkit-fill-available !important;
+    max-height: 100%;
+    position: relative;
+    background-color: white !important;
+  }
+  ion-slide {
+    max-height: 300px;
+  }
+  .circle {
+    width: 33px;
+    height: 33px;
+    margin: auto;
+    border-radius: 25px;
+  }
+  .left {
+    margin-right: 5px;
+    order: 0;
+    flex: 0 1 auto;
+    align-self: auto;
+    position: absolute;
+    left: 0;
+    top: 40%;
+    z-index: 99999;
+    margin-left: 5px;
+  }
+  .right {
+    margin-left: 5px;
+    order: 0;
+    flex: 0 1 auto;
+    align-self: auto;
+    position: absolute;
+    right: 0;
+    top: 40%;
+    z-index: 99999;
+    margin-right: 5px;
+  }
+}
+
+@media only screen and (max-width: 600px) {
+  wm-gallery {
+    webmapp-image-modal {
+      width: 100%;
+      max-width: fit-content;
+    }
+  }
+}
