+++ src/app/components/common/wm-common.module.ts
@@ -8,6 +8,7 @@ import {PoiBoxComponent} from './poi-box/poi-box.component';
 import {RelatedUrlsComponent} from './related-urls.component';
 import {SearchBoxComponent} from './search-box/search-box.component';
 import {TrackAudioComponent} from './track-audio/track-audio.component';
+import {GalleryComponent} from './gallery/gallery.component';
 
 const components = [
   RelatedUrlsComponent,
@@ -16,6 +17,7 @@ const components = [
   LayerBoxComponent,
   PoiBoxComponent,
   TrackAudioComponent,
+  GalleryComponent,
 ];
 @NgModule({
   declarations: components,
