--- src/app/components/track-details/gallery/gallery.component.html
@@ -1,11 +1,0 @@
-<img *ngFor="let item of gallery" [src]="item.url" (click)="currentImage$.next(item)" />
-
-<webmapp-image-modal
-  *ngIf="(currentImage$|async) as currentImage"
-  [image]="currentImage.url !=null? currentImage.url:defaultPhotoPath"
-  [showArrows]="true"
-  [caption]="currentImage.caption"
-  (closeEVT)="close()"
-  (nextEVT)="next()"
-  (prevEVT)="prev()"
-></webmapp-image-modal>
