--- src/app/components/track-details/gallery/gallery.component.scss
@@ -1,33 +1,0 @@
-webmapp-gallery {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: space-around;
-  align-content: flex-start;
-  align-items: flex-start;
-  > img {
-    flex: 0 0 auto;
-    align-self: auto;
-    border-radius: 15px;
-    padding: 10px;
-    object-fit: cover;
-    height: 150px;
-    width: 50%;
-    border-radius: 15px;
-    cursor: pointer;
-  }
-  webmapp-image-modal {
-    width: 50%;
-    padding: 10px;
-    max-width: fit-content;
-  }
-}
-
-@media only screen and (max-width: 600px) {
-  webmapp-gallery {
-    webmapp-image-modal {
-      width: 100%;
-      max-width: fit-content;
-    }
-  }
-}
