+++ src/theme/33.css
@@ -29,7 +29,7 @@
 webmapp-track-description {
   margin-top: -15px;
 }
-webmapp-gallery {
+wm-gallery {
   order: 2;
 }
 webmapp-track-elevation-chart {
